+++ Caller-ID/release/output.json
@@ -1,0 +1,1 @@
+[{"outputType":{"type":"APK"},"apkData":{"type":"MAIN","splits":[],"versionCode":1,"versionName":"1.0","enabled":true,"outputFile":"Caller-ID-release.apk","fullName":"release","baseName":"release","dirName":""},"path":"Caller-ID-release.apk","properties":{}}]
