+++ Caller-ID/src/main/java/com/example/Caller_ID/ui/spamProtection/SpamAdapter.java
@@ -80,7 +80,6 @@ public class SpamAdapter extends RecyclerView.Adapter<SpamAdapter.ViewHolder> im
                         }
                     }
                     spamersListFiltered = filteredSpamers;
-
                 }
                 FilterResults filterResults = new FilterResults();
                 filterResults.values = spamersListFiltered;
